Update glean tests to 2a932dcf5d3aa513ecb65b51555408fb4783ba1a+fixes
This still contains one of my two blendFunc performance fixes that
were never applied.  The other one got reimplemented by probing less
instead of just drawing fewer pixels, for whatever reason.

+++ tests/glean/treadpixperf.cpp
@@ -207,7 +207,7 @@ ReadpixPerfTest::runNonPBOtest(int formatNum, GLsizei width, GLsizei height,
 
 	delete buffer;
 
-	double rate = width * height * iter / elapsedTime / 1000000.0;
+	double rate = static_cast<double>(width) * height * iter / elapsedTime / 1000000.0;
 	return rate;
 }
 
@@ -299,7 +299,7 @@ ReadpixPerfTest::runPBOtest(int formatNum, GLsizei width, GLsizei height,
 	delete b;
 #endif
 
-	double rate = width * height * iter / elapsedTime / 1000000.0;
+	double rate = static_cast<double>(width) * height * iter / elapsedTime / 1000000.0;
 	return rate;
 #else
 	return 0.0;
